Moved icons to respective folder
