+++ packages/seatmaps-client/src/icons/MissingSeatMapLogo.tsx
